Broke up dashboard into multiple components
