+++ Frontend/src/Components/Dashboard/Scores.jsx
@@ -1,0 +1,10 @@
+import { Card } from "flowbite-react";
+export default function Scores () {
+    return (
+      <>
+        <Card className="card-container items-center justify-center text-2xl border-2">
+          Scores
+        </Card>
+      </>
+    );
+}
