+++ Frontend/src/Components/Dashboard/Stats.jsx
@@ -1,0 +1,11 @@
+import { Card } from "flowbite-react";
+
+export default function Stats () {
+    return (
+      <>
+        <Card className="card-container items-center justify-center text-2xl border-2">
+          Stats
+        </Card>
+      </>
+    );
+}
