--- Frontend/src/Components/Header.jsx
@@ -1,10 +1,0 @@
-import { Button } from "flowbite-react";
-export default function Header () {
-    return (
-      <>
-        <Button className="btn-style-1">Log out</Button>
-        <h1 className="title-style">Dashboard</h1>
-        <Button className="btn-style-1">Profile</Button>
-      </>
-    );
-}
