--- Frontend/src/Components/Stats.jsx
@@ -1,9 +1,0 @@
-import { Card } from "flowbite-react";
-
-export default function Stats () {
-    return (
-      <>
-        
-      </>
-    );
-}
